requirements.txt
langchain
langchain-community
langchain-openai
pinecone-client
langchain-groq
protoc_gen_openapiv2
langchain-pinecone
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
sentence-transformers==2.2.2